patch №1
bam

+++ mods/anomaly/code/anomaly_deleting.dm
@@ -5,9 +5,17 @@
 	delete_anomaly()
 
 /obj/anomaly/proc/delete_anomaly()
+	if(!detectable_effect_range)
+		return
 	if(multitile)
 		for(var/obj/anomaly/part in list_of_parts)
+			if(AnomaliesAmmountInTurf(part.loc) == 1)
+				var/turf/part_turf = get_turf(part)
+				part_turf.in_anomaly_effect_range = FALSE
 			qdel(part)
+	if(AnomaliesAmmountInTurf(src.loc) == 1)
+		var/turf/part_turf = get_turf(src)
+		part_turf.in_anomaly_effect_range = FALSE
 	qdel(src)
 
 /obj/anomaly/part/shuttle_land_on()
